Add a `container-build` Makefile target
This new target will build the pf9-clusteradm binary inside a
Linux-based container (Debian 9 at the time of writing) using
the latest golang.

This will mount the current working directory as a volume to the
container, set the volume as the working directory and run `make`.

This target will be used for TeamCity builds.