Update APIEndpoints
Add master: If VIP is configured, update APIEndpoints with the VIP. Otherwise,
update APIEndpoints with SSHConfig.Host (not correct, but consistent with
current implementation).

Delete master: If VIP is configured and the last master is being deleted,
remove the VIP from APIEndpoints. If VIP is configured and there are other
masters, do not remove the VIP from the APIEndpoints. Otherwise, remove the
endpoint that matches SSHConfig.Host.

Fixes #25